Initial Development of Input Field
Added function to generate fock pulses (in a new file) and the input field generator to implement the input field. As per previous email, we can discuss if this should be user side or strictly developer side used in the time evolution functions. Also included some simple pulse envelope functions.

Added an example for the pulsed input.

Also created a simple function to take the expectation of a list of OC normalized bins with a list of operators to get list of time dependent expectation values. I think something like this to be the main way to take expectation values would be nice as it allows a lot of user flexibility (what expectation values they want to take can be whatever operator they define, no longer any system dependence on the expectation function). We can discuss this further if you want.