Simplified _svd_tensors function interface
Removed the first two parameters of the svd function and moved appropriate shaping to be done by the function. I haven't seen anywhere where custom left/right shapes are needed, but feel free to reject if there is a use case in which case it would be important to have.