Fock Pulse Feature Added
As discussed, included code and an example for a non-trivial input photonic field. Feel free to use all or none of the code in the branch (accept or reject). We can also use it as a base and create a wrapper function around it if we want to limit it's usage in some way while using the underlying function still.

Also added a function to take a list of OC normalized bins and a list of operators and computed a list of time dependent expectation values of operators. I think moving to taking expectation values like this from a user side would be good, as it gives the user a lot more versatility in the operators/expectation values they take.